@@ -51,17 +51,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/cdk": "~13.3.0",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/elements": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/material": "~13.3.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
+    "@angular/animations": "^14.2.12",
+    "@angular/cdk": "^13.3.0",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/elements": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/material": "^13.3.0",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "@senzing/rest-api-client-ng": "^5.1.0-beta.1",
     "acorn": "^8.6.0",
     "ajv": "^6.12.6",
@@ -76,12 +76,12 @@
     "zone.js": "~0.11.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.0",
-    "@angular-devkit/core": "^13.3.0",
-    "@angular-devkit/schematics": "^13.3.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "~13.3.0",
-    "@angular/language-service": "^13.3.0",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular-devkit/core": "^14.2.10",
+    "@angular-devkit/schematics": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
+    "@angular/language-service": "^14.2.12",
     "@compodoc/compodoc": "^1.1.18",
     "@types/d3": "^5.9.1",
     "@types/jasmine": "~3.5.0",
@@ -100,14 +100,14 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~1.7.0",
     "karma-summary-reporter": "~3.1.1",
-    "ng-packagr": "^13.3.0",
+    "ng-packagr": "^14.2.2",
     "ngx-webstorage-service": "^5.0.0",
     "protractor": "~7.0.0",
     "rimraf": "^3.0.2",
     "sass-burger": "^1.3.1",
     "scss-bundle": "^3.1.2",
     "ts-node": "~10.9.1",
-    "typescript": "~4.4.4",
+    "typescript": "~4.8.4",
     "wait-on": "^6.0.1"
   },
   "optionalDependencies": {
