@@ -102,125 +102,104 @@ const SzRestConfigurationInjector = new InjectionToken<SzRestConfiguration>("SzR
  * Add to your applications module imports array.
  */
 @NgModule({
-  declarations: [
-    SzAlertMessageDialog,
-    SzConfigurationAboutComponent,
-    SzConfigurationComponent,
-    SzEntityDetailComponent,
-    SzEntityDetailGraphControlComponent,
-    SzEntityDetailGraphComponent,
-    SzEntityDetailGraphFilterComponent,
-    SzEntityDetailHeaderComponent,
-    SzEntityDetailHeaderContentComponent,
-    SzEntityDetailSectionSummaryComponent,
-    SzEntityDetailSectionHeaderComponent,
-    SzEntityDetailSectionCollapsibleCardComponent,
-    SzEntityDetailsSectionComponent,
-    SzEntityMatchPillComponent,
-    SzEntityRecordCardComponent,
-    SzEntityRecordViewerComponent,
-    SzEntityRecordCardHeaderComponent,
-    SzEntityRecordCardContentComponent,
-    SzGraphControlComponent,
-    SzGraphComponent,
-    SzGraphFilterComponent,
-    SzMultiSelectButtonComponent,
-    SzPoweredByComponent,
-    SzPreferencesComponent,
-    SzPrefDictComponent,
-    SzRelationshipNetworkComponent,
-    SzRelationshipNetworkInputComponent,
-    SzRelationshipNetworkLookupComponent,
-    SzRelationshipPathComponent,
-    SzSearchComponent,
-    SzSearchByIdComponent,
-    SzSearchIdentifiersPickerDialogComponent,
-    SzSearchIdentifiersPickerSheetComponent,
-    SzSearchResultsComponent,
-    SzSearchResultCardComponent,
-    SzSearchResultCardContentComponent,
-    SzSearchResultCardHeaderComponent,
-    SzStandaloneGraphComponent,
-    SzWhyEntitiesComparisonComponent,
-    SzWhyEntityComponent,
-    SzWhyEntitiesDialog,
-    SzWhyEntityDialog
-  ],
-  imports: [
-    CommonModule,
-    DragDropModule,
-    HttpClientModule,
-    FormsModule,
-    ReactiveFormsModule,
-    LayoutModule,
-    NgxJsonViewerModule,
-    ApiModule,
-    SzSdkMaterialModule
-  ],
-  exports: [
-    SzConfigurationComponent,
-    SzConfigurationAboutComponent,
-    SzEntityDetailGraphComponent,
-    SzEntityDetailComponent,
-    SzEntityDetailGraphControlComponent,
-    SzEntityDetailGraphFilterComponent,
-    SzEntityRecordViewerComponent,
-    SzGraphComponent,
-    SzGraphControlComponent,
-    SzGraphFilterComponent,
-    SzPoweredByComponent,
-    SzPreferencesComponent,
-    SzRelationshipNetworkComponent,
-    SzRelationshipNetworkInputComponent,
-    SzRelationshipNetworkLookupComponent,
-    SzRelationshipPathComponent,
-    SzSearchComponent,
-    SzSearchByIdComponent,
-    SzSearchResultsComponent,
-    SzSearchResultCardComponent,
-    SzStandaloneGraphComponent,
-    SzWhyEntitiesComparisonComponent,
-    SzWhyEntityComponent,
-    SzWhyEntitiesDialog,
-    SzWhyEntityDialog
-  ],
-  /** for components being exported as web components */
-  entryComponents: [
-    SzAlertMessageDialog,
-    SzConfigurationComponent,
-    SzConfigurationAboutComponent,
-    SzEntityDetailComponent,
-    SzEntityDetailGraphComponent,
-    SzEntityRecordViewerComponent,
-    SzPoweredByComponent,
-    SzPreferencesComponent,
-    SzSearchComponent,
-    SzSearchByIdComponent,
-    SzSearchIdentifiersPickerDialogComponent,
-    SzSearchIdentifiersPickerSheetComponent,
-    SzSearchResultsComponent,
-    SzStandaloneGraphComponent,
-    SzWhyEntitiesComparisonComponent,
-    SzWhyEntityComponent,
-    SzWhyEntitiesDialog,
-    SzWhyEntityDialog
-  ],
-  providers: [
-    SzMessageBundleService,
-    SzAdminService,
-    SzBulkDataService,
-    SzConfigurationService,
-    SzCSSClassService,
-    SzDataSourcesService,
-    SzFoliosService,
-    SzPrefsService,
-    SzSearchService,
-    HttpClient,
-    TitleCasePipe,
-    SzUIEventService,
-    SzPdfUtilService,
-    Location
-  ]
+    declarations: [
+        SzAlertMessageDialog,
+        SzConfigurationAboutComponent,
+        SzConfigurationComponent,
+        SzEntityDetailComponent,
+        SzEntityDetailGraphControlComponent,
+        SzEntityDetailGraphComponent,
+        SzEntityDetailGraphFilterComponent,
+        SzEntityDetailHeaderComponent,
+        SzEntityDetailHeaderContentComponent,
+        SzEntityDetailSectionSummaryComponent,
+        SzEntityDetailSectionHeaderComponent,
+        SzEntityDetailSectionCollapsibleCardComponent,
+        SzEntityDetailsSectionComponent,
+        SzEntityMatchPillComponent,
+        SzEntityRecordCardComponent,
+        SzEntityRecordViewerComponent,
+        SzEntityRecordCardHeaderComponent,
+        SzEntityRecordCardContentComponent,
+        SzGraphControlComponent,
+        SzGraphComponent,
+        SzGraphFilterComponent,
+        SzMultiSelectButtonComponent,
+        SzPoweredByComponent,
+        SzPreferencesComponent,
+        SzPrefDictComponent,
+        SzRelationshipNetworkComponent,
+        SzRelationshipNetworkInputComponent,
+        SzRelationshipNetworkLookupComponent,
+        SzRelationshipPathComponent,
+        SzSearchComponent,
+        SzSearchByIdComponent,
+        SzSearchIdentifiersPickerDialogComponent,
+        SzSearchIdentifiersPickerSheetComponent,
+        SzSearchResultsComponent,
+        SzSearchResultCardComponent,
+        SzSearchResultCardContentComponent,
+        SzSearchResultCardHeaderComponent,
+        SzStandaloneGraphComponent,
+        SzWhyEntitiesComparisonComponent,
+        SzWhyEntityComponent,
+        SzWhyEntitiesDialog,
+        SzWhyEntityDialog
+    ],
+    imports: [
+        CommonModule,
+        DragDropModule,
+        HttpClientModule,
+        FormsModule,
+        ReactiveFormsModule,
+        LayoutModule,
+        NgxJsonViewerModule,
+        ApiModule,
+        SzSdkMaterialModule
+    ],
+    exports: [
+        SzConfigurationComponent,
+        SzConfigurationAboutComponent,
+        SzEntityDetailGraphComponent,
+        SzEntityDetailComponent,
+        SzEntityDetailGraphControlComponent,
+        SzEntityDetailGraphFilterComponent,
+        SzEntityRecordViewerComponent,
+        SzGraphComponent,
+        SzGraphControlComponent,
+        SzGraphFilterComponent,
+        SzPoweredByComponent,
+        SzPreferencesComponent,
+        SzRelationshipNetworkComponent,
+        SzRelationshipNetworkInputComponent,
+        SzRelationshipNetworkLookupComponent,
+        SzRelationshipPathComponent,
+        SzSearchComponent,
+        SzSearchByIdComponent,
+        SzSearchResultsComponent,
+        SzSearchResultCardComponent,
+        SzStandaloneGraphComponent,
+        SzWhyEntitiesComparisonComponent,
+        SzWhyEntityComponent,
+        SzWhyEntitiesDialog,
+        SzWhyEntityDialog
+    ],
+    providers: [
+        SzMessageBundleService,
+        SzAdminService,
+        SzBulkDataService,
+        SzConfigurationService,
+        SzCSSClassService,
+        SzDataSourcesService,
+        SzFoliosService,
+        SzPrefsService,
+        SzSearchService,
+        HttpClient,
+        TitleCasePipe,
+        SzUIEventService,
+        SzPdfUtilService,
+        Location
+    ]
 })
 export class SenzingSdkModule {
   /**
